+++ shaders/import/dhFarPlane.glsl
@@ -1,0 +1,5 @@
+
+#define import_dhFarPlane
+#ifdef FIRST_PASS
+float dhFarPlane = 0.0;
+#endif