+++ shaders/import/dhNearPlane.glsl
@@ -1,0 +1,5 @@
+
+#define import_dhNearPlane
+#ifdef FIRST_PASS
+float dhNearPlane = 0.0;
+#endif