+++ shaders/import/dhProjectionInverse.glsl
@@ -1,0 +1,5 @@
+
+#define import_dhProjectionInverse
+#ifdef FIRST_PASS
+mat4 dhProjectionInverse = mat4(0.0);
+#endif