+++ shaders/style_cartoon.glsl
@@ -732,9 +732,9 @@
 #undef SUNRAYS_INC_AMOUNT
 										#define SUNRAYS_INC_AMOUNT 1.5
 #endif
-#if SUNRAYS_ENTER_AMOUNT == -1
-#undef SUNRAYS_ENTER_AMOUNT
-										#define SUNRAYS_ENTER_AMOUNT 3.0
+#if SUNRAYS_DEC_AMOUNT == -1
+#undef SUNRAYS_DEC_AMOUNT
+										#define SUNRAYS_DEC_AMOUNT 0.5
 #endif